Clean up, clean up, everybody do your share
