+++ Assets/Animations/StarPowerParticles.controller.meta
