--- Assets/Script/PlayMode/Animations.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 76e634e8478c6fe498be2aecb7749133
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
