--- Assets/Script/PlayMode/TrckAnimations.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 21698035b4764784a8a42c708d3a3482
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
