--- Assets/TrackAnimations.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ca69246557021364ea03fa3802fcf40a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
